+++ mbo/hash/hash_simple.h
@@ -13,8 +13,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef MBO_HASH_SIMPLE_HASH_H_
-#define MBO_HASH_SIMPLE_HASH_H_
+#ifndef MBO_HASH_HASH_SIMPLE_H_
+#define MBO_HASH_HASH_SIMPLE_H_
 
 // IWYU pragma: private, include "mbo/hash/hash.h"
 
@@ -23,11 +23,14 @@
 #include <string_view>
 #include <type_traits>
 
+namespace mbo::hash::simple {
+
+// NOLINTBEGIN(*-identifier-naming,*-magic-numbers,*-constant-array-index,*-pointer-arithmetic,*-signed-bitwise)
+
 // Namespace `mbo::hash::simple` offers a `GetHash` a constexpr safe simple hash function.
-namespace mbo::hash::hash_internal {
+namespace hash_internal {
 
 inline constexpr uint64_t GetSimpleHash(std::string_view data) {
-  // NOLINTBEGIN(*-magic-numbers)
   constexpr uint64_t kArbitrary = 5'008'709'998'333'326'415ULL;
   constexpr uint64_t kPrimeNum10k = 104'729ULL;
   if (data.empty()) {
@@ -52,14 +55,14 @@ inline constexpr uint64_t GetSimpleHash(std::string_view data) {
       add += data[pos++] << 8U;
       add += data[pos++] << 16U;
       add += data[pos++] << 24U;
-      result = result * 6'571 ^ ((17 * add + (add >> 16U)) ^ (result >> 32U));
+      result = (result * 6'571) ^ ((17 * add + (add >> 16U)) ^ (result >> 32U));
       len -= 4;
     }
     uint64_t add = 0;
     switch (len) {
       case 3: add += data[pos + 2] << 16U;
       case 2: add += data[pos + 1] << 8U;
-      case 1: add += data[pos]; return result * 193 + kPrimeNum10k * add;
+      case 1: add += data[pos]; return (result * 193) + (kPrimeNum10k * add);
       default: break;
     }
     return result;
@@ -70,70 +73,41 @@ inline constexpr uint64_t GetSimpleHash(std::string_view data) {
     uint64_t add = 0;
     while (str < end) {
       std::memcpy(&add, str, 4);
-      result = result * 6'571 ^ ((17 * add + (add >> 16U)) ^ (result >> 32U));
-      str += 4;  // NOLINT(*-pointer-arithmetic)
+      result = (result * 6'571) ^ ((17 * add + (add >> 16U)) ^ (result >> 32U));
+      str += 4;
     }
     switch (data.length() % 4) {  // NOLINT(*-default-case)
       case 3:
         add = 0;
         std::memcpy(&add, str, 3);
-        return result * 193 + kPrimeNum10k * add;
+        return (result * 193) + (kPrimeNum10k * add);
       case 2:
         add = 0;
         std::memcpy(&add, str, 2);
-        return result * 193 + kPrimeNum10k * add;
+        return (result * 193) + (kPrimeNum10k * add);
       case 1:
         add = 0;
         std::memcpy(&add, str, 1);
-        return result * 193 + kPrimeNum10k * add;
+        return (result * 193) + (kPrimeNum10k * add);
       case 0: return result;
     }
     return result;
   }
-  // NOLINTEND(*-magic-numbers)
 }
 
-}  // namespace mbo::hash::hash_internal
+}  // namespace hash_internal
 
-namespace mbo::hash::simple {
-
-// NOLINTBEGIN(*-macro-usage)
-#define MBO_CONSTANT_P(expression, fallback) (fallback)
-#ifdef __has_builtin
-# if __has_builtin(__builtin_constant_p)
-#  undef MBO_CONSTANT_P
-#  define MBO_CONSTANT_P(expression, fallback) __builtin_constant_p(expression)
-# endif
-#endif
-// NOLINTEND(*-macro-usage)
-
-// A simple constexpr safe hash function.
-//
-// This function uses an optimized implementation at run-time and a constexpr safe implementation
-// that yields the exact same values for compile-time usage. It uses compiler provided macros as a
-// seed generator. This means that the seed **may** vary from run to run, but there is no guarantee
-// as a dev environment (e.g. Bazel) may set these macros always to the same value.
-inline constexpr uint64_t GetHash(std::string_view data) {
-  using mbo::hash::hash_internal::GetSimpleHash;
-  // In theory this should be a random number or at least an arbitray number that
-  // changes on every program start. Howeverm this number must be constexpr.
-  constexpr uint64_t kNotSoRandom =  //
-      5'008'709'998'333'326'415ULL
-#if defined(__DATE__)
-      ^ (MBO_CONSTANT_P(__DATE__, 1) == 0 ? 0 : GetSimpleHash(std::string_view(__DATE__)))
-#endif
-#if defined(__TIME__)
-      ^ (MBO_CONSTANT_P(__TIME__, 1) == 0 ? 0 : GetSimpleHash(std::string_view(__TIME__)))
-#endif
-#if defined(__TIMESTAMP__)
-      ^ (MBO_CONSTANT_P(__TIMESTAMP__, 1) == 0 ? 0 : GetSimpleHash(std::string_view(__TIMESTAMP__)))
-#endif
-      ;
-  return GetSimpleHash(data) ^ kNotSoRandom;
+inline constexpr uint64_t GetHash64(std::string_view data) {
+  return hash_internal::GetSimpleHash(data);
 }
 
-#undef MBO_CONSTANT_P  // Be gone.
+// Deprecated in version 0.13.0
+[[deprecated("Use ::mbo::hash::GetHash64() instead.")]] inline constexpr uint64_t GetHash(std::string_view data) {
+  return hash_internal::GetSimpleHash(data);
+}
+
+// NOLINTEND(*-identifier-naming,*-magic-numbers,*-constant-array-index,*-pointer-arithmetic,*-signed-bitwise)
 
 }  // namespace mbo::hash::simple
 
-#endif  // MBO_HASH_SIMPLE_HASH_H_
+#endif  // MBO_HASH_HASH_SIMPLE_H_